Started adding the js functionality to the game
The link to the js file was broken, so that was fixed.
Implemented the first lines of code to the js file, making it so when the player chooses a name, it shows the rest of the game ui and changes it's name in the score.

The images of the rock, paper and scissors has been transformed into buttons.